Add enable_tls option to kafka_consumer
